Move stringManipulationService. Note this won't quite run. Jussa second.
