+++ helpers/stringManipulation.js
